+++ include/ff/detail/ordered_stream_filter.h
@@ -23,51 +23,14 @@
 
 #ifdef GRPPI_FF
 
+#include "nodes.h"
+
 #include <ff/farm.hpp>
 #include <ff/allocator.hpp>
 
 namespace grppi {
 
 namespace detail_ff {
-
-constexpr size_t ordered_filtered_value = (ff::FF_EOS-0x11);
-
-template <typename Item, typename Predicate>
-class ordered_filter_worker : public ff::ff_node_t<Item> {
-public:
-  ordered_filter_worker(Predicate && predicate) :
-      predicate_{predicate}
-  {}
-
-  Item * svc(Item * p_item) {
-    if (predicate_(*p_item)) {
-      return p_item;
-    }
-    else {
-      p_item->~Item();
-      ff::ff_free(p_item);
-      return reinterpret_cast<Item*>(ordered_filtered_value);
-    }
-  }
-
-private:
-  Predicate predicate_;
-};
-
-template <typename Item>
-class ordered_filter_collector : public ff::ff_node_t<Item> {
-public:
-  ordered_filter_collector() = default;
-
-  Item * svc(Item * p_item) {
-    if (p_item == reinterpret_cast<Item*>(ordered_filtered_value)) {
-      return this->GO_ON;
-    }
-    else {
-      return p_item;
-    }
-  }
-};
 
 template <typename Item, typename Filter>
 class ordered_stream_filter : public ff::ff_ofarm {
