+++ include/ff/detail/unordered_stream_filter.h
@@ -23,63 +23,14 @@
 
 #ifdef GRPPI_FF
 
+#include "nodes.h"
+
 #include <ff/farm.hpp>
 #include <ff/allocator.hpp>
 
 namespace grppi {
 
 namespace detail_ff {
-
-constexpr size_t unordered_filtered_value = (ff::FF_EOS-0x11);
-
-template <typename Item, typename Predicate>
-class unordered_filter_worker : public ff::ff_node_t<Item> {
-public:
-  unordered_filter_worker(Predicate && predicate) :
-      predicate_{predicate}
-  {}
-
-  Item * svc(Item * p_item) {
-    if (predicate_(*p_item)) {
-      return p_item;
-    }
-    else {
-      p_item->~Item();
-      ff::ff_free(p_item);
-      return reinterpret_cast<Item*>(unordered_filtered_value);
-    }
-  }
-
-private:
-  Predicate predicate_;
-};
-
-
-template <typename Item>
-class unordered_filter_collector : public ff::ff_node_t<Item> {
-public:
-  unordered_filter_collector() = default;
-
-  Item * svc(Item * p_item) {
-    static_assert(sizeof(std::size_t) == sizeof(std::uintptr_t));
-    if (p_item == reinterpret_cast<Item*>(unordered_filtered_value)) {
-      return this->GO_ON;
-    }
-    else {
-      return p_item;
-    }
-  }
-};
-template <typename Item>
-class unordered_filter_emitter : public ff::ff_node_t<Item> {
-public:
-  unordered_filter_emitter() = default;
-
-  Item * svc(Item * p_item) {
-    return p_item;
-  }
-};
-
 
 template <typename Item, typename Filter>
 class unordered_stream_filter : public ff::ff_ofarm {
